backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=10.0.0
python-dotenv==1.0.1
google-generativeai>=0.3.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
drf-spectacular==0.27.0
markdown==3.6
django-filter==24.1
celery==5.3.6
redis==5.0.1

